BMP transparency optimizations
Transparency, whether through PNG or a color key on BMP, adds a lot of overhead. For BMP, we can disable transparency for sprites that don't need it. I disabled it for background tiles and even split the main grid into three sprites, the biggest of which doesn't need transparency. This greatly improves performance on 3DS, which doesn't support hardware rendering.
Also re-added "Exit Game and Apply Changes" and all resolution options to 3DS Video menu